+++ src/widget/mod.rs
@@ -3,4 +3,5 @@ pub mod combo_box;
 pub mod image;
 pub mod misc;
 pub mod progress_bar;
+pub mod selectable;
 pub mod text;