+++ CHANGELOG.markdown
@@ -16,6 +16,7 @@
 - Redesigned child window API (previously known as child frame)
 - Redesigned image / image button API
 - Redesigned combo box API
+- Redesigned selectable API
 - Updated layout API
 - Renderer errors implement std::error::Error
 - Glium renderer re-exports imgui and glium